New styles for donation payment summary
