+++ src/components/pages/donation_form/summary_content/PaymentSummaryCompanyWithContact.vue
@@ -1,10 +1,10 @@
 <template>
-	<div class="payment-summary" v-html="getSummary()"/>
+	<span v-html="getSummary()"/>
 </template>
 
 <script>
 export default {
-	name: 'PaymentSummaryPrivate',
+	name: 'PaymentSummaryCompany',
 	props: [
 		'address',
 		'interval',
@@ -12,7 +12,6 @@ export default {
 		'paymentType',
 		'country',
 		'languageItem',
-		'salutation',
 	],
 	methods: {
 		getSummary: function () {
@@ -22,7 +21,7 @@ export default {
 					interval: this.$props.interval,
 					formattedAmount: this.$props.formattedAmount,
 					paymentType: this.$props.paymentType,
-					personType: this.$t( 'donation_confirmation_topbox_donor_type_person' ),
+					personType: this.$t( 'donation_confirmation_topbox_donor_type_company' ),
 					address: this.addressString(),
 					email: this.email(),
 				}
@@ -34,27 +33,19 @@ export default {
 			}
 			return this.address.email;
 		},
-		salutationDisplay: function () {
-			if ( !this.salutation || this.salutation === '' ) {
-				return '';
-			}
-			return this.salutation + ' ';
-		},
 		addressString: function () {
 			if ( !this.canRenderAddress() ) {
 				return this.$t( 'donation_confirmation_review_address_missing' );
 			}
 			return [
-				this.salutationDisplay() + this.$props.address.fullName,
+				this.$props.address.fullName,
 				this.$props.address.streetAddress,
 				this.$props.address.postalCode + ' ' + this.$props.address.city,
 				this.$props.country,
 			].join( ', ' );
 		},
 		canRenderAddress: function () {
-			return this.$props.address.salutation
-				&& this.$props.address.firstName
-				&& this.$props.address.lastName
+			return this.$props.address.fullName
 				&& this.$props.address.streetAddress
 				&& this.$props.address.postalCode
 				&& this.$props.address.city;
