+++ src/components/pages/donation_form/summary_content/PaymentSummaryEmail.vue
@@ -1,10 +1,10 @@
 <template>
-	<div class="payment-summary" v-html="getSummary()"></div>
+	<span v-html="getSummary()"/>
 </template>
 
 <script>
 export default {
-	name: 'PaymentSummaryCompany',
+	name: 'PaymentSummaryEmail',
 	props: [
 		'address',
 		'interval',
@@ -12,6 +12,7 @@ export default {
 		'paymentType',
 		'country',
 		'languageItem',
+		'salutation',
 	],
 	methods: {
 		getSummary: function () {
@@ -21,7 +22,7 @@ export default {
 					interval: this.$props.interval,
 					formattedAmount: this.$props.formattedAmount,
 					paymentType: this.$props.paymentType,
-					personType: this.$t( 'donation_confirmation_topbox_donor_type_company' ),
+					personType: this.$t( 'donation_confirmation_topbox_donor_type_email' ),
 					address: this.addressString(),
 					email: this.email(),
 				}
@@ -33,22 +34,22 @@ export default {
 			}
 			return this.address.email;
 		},
+		salutationDisplay: function () {
+			if ( !this.salutation || this.salutation === '' ) {
+				return '';
+			}
+			return this.salutation + ' ';
+		},
 		addressString: function () {
 			if ( !this.canRenderAddress() ) {
 				return this.$t( 'donation_confirmation_review_address_missing' );
 			}
-			return [
-				this.$props.address.fullName,
-				this.$props.address.streetAddress,
-				this.$props.address.postalCode + ' ' + this.$props.address.city,
-				this.$props.country,
-			].join( ', ' );
+			return this.salutationDisplay() + this.$props.address.fullName;
 		},
 		canRenderAddress: function () {
-			return this.$props.address.fullName
-				&& this.$props.address.streetAddress
-				&& this.$props.address.postalCode
-				&& this.$props.address.city;
+			return this.$props.address.salutation
+				&& this.$props.address.firstName
+				&& this.$props.address.lastName;
 		},
 	},
 };
