+++ src/components/shared/form_elements/CityAutocompleteField.vue
@@ -32,9 +32,9 @@
 
 <script setup lang="ts">
 import { computed, inject, onMounted, ref, watch } from 'vue';
-import { useCitiesResource } from '@src/components/shared/form_fields/useCitiesResource';
+import { useCitiesResource } from '@src/components/shared/form_elements/useCitiesResource';
 import { CityAutocompleteResource, NullCityAutocompleteResource } from '@src/CityAutocompleteResource';
-import { useCityAutocompleteEvents } from '@src/components/shared/form_fields/useCityAutocompleteEvents';
+import { useCityAutocompleteEvents } from '@src/components/shared/form_elements/useCityAutocompleteEvents';
 import TextFormInput from '@src/components/shared/form_inputs/TextFormInput.vue';
 
 interface Props {
