+++ src/components/shared/form_elements/CountryAutocompleteField.vue
@@ -38,7 +38,7 @@ import { useCountryInput } from '@src/components/shared/legacy_form_inputs/useCo
 import { useFilteredCountries } from '@src/components/shared/legacy_form_inputs/useFilteredCountries';
 import { Country } from '@src/view_models/Country';
 import TextFormInput from '@src/components/shared/form_inputs/TextFormInput.vue';
-import { useCountryAutocompleteEvents } from '@src/components/shared/form_fields/useCountryAutocompleteEvents';
+import { useCountryAutocompleteEvents } from '@src/components/shared/form_elements/useCountryAutocompleteEvents';
 
 interface Props {
 	modelValue: string;
