+++ src/components/shared/form_elements/EmailField.vue
@@ -22,8 +22,8 @@
 </template>
 
 <script setup lang="ts">
-import { useFieldModel } from '@src/components/shared/form_fields/useFieldModel';
-import { useSuggestedEmailProvider } from '@src/components/shared/form_fields/useSuggestedEmailProvider';
+import { useFieldModel } from '@src/components/shared/form_elements/useFieldModel';
+import { useSuggestedEmailProvider } from '@src/components/shared/form_elements/useSuggestedEmailProvider';
 import { useMailHostList } from '@src/components/shared/useMailHostList';
 import TextFormInput from '@src/components/shared/form_inputs/TextFormInput.vue';
