+++ src/scss/components/button.scss
@@ -1,0 +1,3 @@
+.form-button {
+
+}
