+++ src/scss/components/donation_form_summary.scss
@@ -1,0 +1,17 @@
+@use '../settings/forms';
+@use '../settings/units';
+@use 'sass:map';
+
+.form-summary {
+	padding: map.get( forms.$form-summary, 'padding' );
+	margin: map.get( forms.$form-summary, 'margin' );
+	border: map.get( forms.$form-summary, 'border' );
+
+	&-content {
+		margin: map.get( forms.$form-summary-content, 'margin' );
+	}
+
+	&-buttons {
+
+	}
+}
