+++ src/scss/custom.scss
@@ -4,6 +4,8 @@
 @import "fonts";
 
 // Import Bulma styles
+// TODO: Look through what Bulma elements are being used and only import styles for them
+//    This will reduce the css file size
 @import "~bulma";
 @import "~bulma-helpers/sass/helpers/spacing/margin-padding";
