+++ src/scss/settings/_colors.scss
@@ -1,3 +1,9 @@
 $primary: #0065a4;
 
+$white: #ffffff;
+$black: #000000;
+
+$gray-lighter: #E5E5E5;
+$gray-light: #E6D4D3;
 $gray-dark: #808080;
+$gray-mid: #B7B7B7;
